refactor: コンポーネント構造の再編成 - app/_componentsからfeaturesへの移行 (#118)
## 変更内容

### 認証関連コンポーネントの移動
- LoginForm, CreateAccountForm等をfeatures/auth/components/forms/へ移動
- ConfirmEmailClientをfeatures/auth/components/email/へ移動

### ダッシュボード関連の移動と整理
- app/(dashboard)/_componentsをfeatures/running/components/dashboard/へ移動
- ServerRunningDashboard, DashboardWithCalendar, DashboardStatisticsを配置

### ランニングコンポーネントの機能別整理
- calendar/: ClientRunningCalendar
- charts/: RunningChart, RunningChartWrapper
- forms/: ClientGoalForm, ClientRecordForm, ClientYearlyGoalForm
- statistics/: RecentRecords, RecordDetailModal, StatisticsCards

### その他の変更
- ResponsiveHeaderNavをHeaderNavに統合
- 全page.tsxのインポートパスを更新
- 空の_componentsディレクトリを削除

## 効果
- appディレクトリはルーティングに専念
- featuresディレクトリにビジネスロジックを集約
- 機能別の直感的なディレクトリ構造
