+++ next/src/app/(auth)/create-account/page.tsx
@@ -1,8 +1,7 @@
 import { redirect } from 'next/navigation';
 
+import CreateAccountForm from '@/features/auth/components/forms/CreateAccountForm';
 import { getAuthStatus } from '@/features/auth/lib/server';
-
-import CreateAccountForm from './_components/CreateAccountForm';
 
 export default async function CreateAccountPage() {
   // 既にログイン済みの場合はダッシュボードへリダイレクト
@@ -13,8 +12,8 @@ export default async function CreateAccountPage() {
 
   return (
     <div className="flex min-h-screen items-center justify-center bg-gradient-to-br from-green-100 to-blue-200 px-4 py-8">
-      <div className="w-full max-w-md rounded-xl bg-white p-6 md:p-8 shadow-lg">
-        <h2 className="mb-6 text-center text-2xl md:text-3xl font-bold text-green-700">
+      <div className="w-full max-w-md rounded-xl bg-white p-6 shadow-lg md:p-8">
+        <h2 className="mb-6 text-center text-2xl font-bold text-green-700 md:text-3xl">
           アカウント作成
         </h2>
         <CreateAccountForm />
