+++ next/src/app/(auth)/reset-password/page.tsx
@@ -1,7 +1,7 @@
 import { Metadata } from 'next';
 import { Suspense } from 'react';
 
-import ResetPasswordForm from './_components/ResetPasswordForm';
+import ResetPasswordForm from '@/features/auth/components/forms/ResetPasswordForm';
 
 export const metadata: Metadata = {
   title: '新しいパスワードを設定 | Runmates',
@@ -11,8 +11,8 @@ export const metadata: Metadata = {
 export default function ResetPasswordPage() {
   return (
     <div className="flex min-h-screen items-center justify-center bg-gradient-to-br from-green-100 to-blue-200 px-4 py-8">
-      <div className="w-full max-w-md rounded-xl bg-white p-6 md:p-8 shadow-lg">
-        <h2 className="mb-6 text-center text-2xl md:text-3xl font-bold text-green-700">
+      <div className="w-full max-w-md rounded-xl bg-white p-6 shadow-lg md:p-8">
+        <h2 className="mb-6 text-center text-2xl font-bold text-green-700 md:text-3xl">
           新しいパスワードを設定
         </h2>
         <p className="mb-6 text-center text-sm text-gray-600">
