+++ next/src/features/running/components/charts/RunningChart.tsx
@@ -382,7 +382,7 @@ export default function RunningChart({
   };
 
   return (
-    <div className="rounded-xl bg-white p-4 md:p-6 shadow-lg">
+    <div className="rounded-xl bg-white p-4 shadow-lg md:p-6">
       <div className="h-64 md:h-80">
         <Line
           data={data as ChartData<'line'>}
@@ -393,7 +393,7 @@ export default function RunningChart({
       <div className="mt-4 mb-4 flex items-center justify-between">
         <button
           onClick={goToPreviousMonth}
-          className="flex items-center rounded-lg px-2 py-1 md:px-3 md:py-2 text-sm md:text-base text-gray-600 transition-colors hover:bg-gray-100 hover:text-gray-800"
+          className="flex items-center rounded-lg px-2 py-1 text-sm text-gray-600 transition-colors hover:bg-gray-100 hover:text-gray-800 md:px-3 md:py-2 md:text-base"
         >
           <span className="mr-1">←</span>
           <span className="hidden sm:inline">前の月</span>
@@ -401,14 +401,14 @@ export default function RunningChart({
         </button>
 
         <div className="flex items-center space-x-2">
-          <h3 className="text-base md:text-lg font-bold text-gray-800">
+          <h3 className="text-base font-bold text-gray-800 md:text-lg">
             {viewYear}年{viewMonth + 1}月
           </h3>
         </div>
 
         <button
           onClick={goToNextMonth}
-          className="flex items-center rounded-lg px-2 py-1 md:px-3 md:py-2 text-sm md:text-base text-gray-600 transition-colors hover:bg-gray-100 hover:text-gray-800"
+          className="flex items-center rounded-lg px-2 py-1 text-sm text-gray-600 transition-colors hover:bg-gray-100 hover:text-gray-800 md:px-3 md:py-2 md:text-base"
         >
           <span className="hidden sm:inline">次の月</span>
           <span className="sm:hidden">次月</span>
@@ -417,22 +417,22 @@ export default function RunningChart({
       </div>
 
       {/* 統計表示 */}
-      <div className="grid grid-cols-3 gap-2 md:gap-4 text-xs md:text-sm text-gray-600">
+      <div className="grid grid-cols-3 gap-2 text-xs text-gray-600 md:gap-4 md:text-sm">
         <div className="text-center">
           <div className="font-semibold text-emerald-600">月間走行距離</div>
-          <div className="text-base md:text-lg font-bold">
+          <div className="text-base font-bold md:text-lg">
             {viewMonthCumulative.toFixed(1)} km
           </div>
         </div>
         <div className="text-center">
           <div className="font-semibold text-blue-600">月間目標</div>
-          <div className="text-base md:text-lg font-bold">
+          <div className="text-base font-bold md:text-lg">
             {monthlyGoal ? `${monthlyGoal.toFixed(1)} km` : '未設定'}
           </div>
         </div>
         <div className="text-center">
           <div className="font-semibold text-purple-600">達成率</div>
-          <div className="text-base md:text-lg font-bold">
+          <div className="text-base font-bold md:text-lg">
             {monthlyGoal && monthlyGoal > 0
               ? `${((viewMonthCumulative / monthlyGoal) * 100).toFixed(0)}%`
               : '-'}
