+++ next/src/features/running/components/forms/ClientRecordForm.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { zodResolver } from '@hookform/resolvers/zod';
-import { useCallback, useState, useTransition } from 'react';
+import { useCallback, useEffect, useState, useTransition } from 'react';
 import { useForm } from 'react-hook-form';
 import * as z from 'zod';
 
@@ -16,7 +16,6 @@ import {
 import {
   Form,
   FormControl,
-  FormDescription,
   FormField,
   FormItem,
   FormLabel,
@@ -24,67 +23,79 @@ import {
 } from '@/components/ui/form';
 import { Input } from '@/components/ui/input';
 
-import { updateMonthlyGoal } from '../actions/running-actions';
+import { createRunningRecord } from '../../actions/running-actions';
 
-const monthlyGoalSchema = z.object({
-  distance_goal: z.union([
-    z.number().min(1, '目標距離は1km以上で入力してください'),
-    z.literal('').transform(() => null),
-    z.null(),
+const runningRecordSchema = z.object({
+  date: z.string().min(1, '日付を入力してください'),
+  distance: z.union([
+    z.number().min(0.01, '距離は0より大きい値を入力してください'),
+    z.literal('').transform(() => 0),
   ]),
 });
 
-type MonthlyGoalFormData = {
-  distance_goal: number | '' | null;
+type RunningRecordFormData = {
+  date: string;
+  distance: number | '';
 };
 
-interface ClientGoalFormProps {
-  currentGoal: number | null;
+interface ClientRecordFormProps {
+  selectedDate?: string;
   isOpen: boolean;
   onClose: () => void;
-  showWelcomeMessage?: boolean;
 }
 
-export default function ClientGoalForm({
-  currentGoal,
+export default function ClientRecordForm({
+  selectedDate,
   isOpen,
   onClose,
-  showWelcomeMessage = false,
-}: ClientGoalFormProps) {
-  const form = useForm<MonthlyGoalFormData>({
-    resolver: zodResolver(monthlyGoalSchema),
+}: ClientRecordFormProps) {
+  const form = useForm<RunningRecordFormData>({
+    resolver: zodResolver(runningRecordSchema),
     defaultValues: {
-      distance_goal: '',
+      date: selectedDate || '',
+      distance: '',
     },
   });
+
+  // モーダルが開かれたときに日付を設定
+  useEffect(() => {
+    if (isOpen) {
+      if (selectedDate) {
+        form.setValue('date', selectedDate);
+      } else {
+        form.setValue('date', new Date().toISOString().split('T')[0]);
+      }
+    }
+  }, [isOpen, selectedDate, form]);
 
   const [isPending, startTransition] = useTransition();
   const [error, setError] = useState<string | null>(null);
 
   const handleClose = useCallback(() => {
-    form.reset();
+    form.reset({
+      date: '',
+      distance: '',
+    });
     setError(null);
     if (onClose) {
       onClose();
     }
   }, [form, onClose]);
 
-  const onSubmit = async (data: MonthlyGoalFormData) => {
+  const onSubmit = async (data: RunningRecordFormData) => {
     setError(null);
     const formData = new FormData();
-    const distance = data.distance_goal === '' || data.distance_goal === null ? null : data.distance_goal;
-    if (distance !== null) {
-      formData.append('distance_goal', distance.toString());
-    }
+    formData.append('date', data.date);
+    const distance = data.distance === '' ? 0 : data.distance;
+    formData.append('distance', distance.toString());
 
-    // 非同期処理をトランジションでラップすることで、この処理の間はisPendingがtrueになる
     startTransition(async () => {
-      const result = await updateMonthlyGoal(formData);
+      const result = await createRunningRecord(formData);
       if (result.success) {
         form.reset();
         handleClose();
       } else {
-        setError(result.error || '目標の設定に失敗しました');
+        setError(result.error || '記録の保存に失敗しました');
       }
     });
   };
@@ -94,37 +105,41 @@ export default function ClientGoalForm({
       <DialogContent className="sm:max-w-md">
         <DialogHeader>
           <DialogTitle className="text-xl font-bold text-gray-800">
-            {showWelcomeMessage
-              ? '🎉 ランニングを始めましょう！'
-              : '🎯 今月の目標を変更'}
+            🏃‍♂️ 新しい走行記録
           </DialogTitle>
           <DialogDescription>
-            {showWelcomeMessage
-              ? '初めての目標を設定して、健康的な習慣を始めましょう。'
-              : '今月の走行距離の目標を変更します。'}
+            本日の走行距離を記録しましょう。
           </DialogDescription>
         </DialogHeader>
-
-        {showWelcomeMessage && (
-          <p className="mb-4 text-gray-600">
-            まずは今月の走行距離目標を設定して、モチベーションを高めましょう！
-          </p>
-        )}
 
         <Form {...form}>
           <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-4">
             <FormField
               control={form.control}
-              name="distance_goal"
+              name="date"
               render={({ field }) => (
                 <FormItem>
-                  <FormLabel>目標距離 (km)</FormLabel>
+                  <FormLabel>日付</FormLabel>
+                  <FormControl>
+                    <Input type="date" {...field} />
+                  </FormControl>
+                  <FormMessage />
+                </FormItem>
+              )}
+            />
+
+            <FormField
+              control={form.control}
+              name="distance"
+              render={({ field }) => (
+                <FormItem>
+                  <FormLabel>距離 (km)</FormLabel>
                   <FormControl>
                     <Input
                       type="number"
                       step="0.1"
-                      min="1"
-                      placeholder="50.0"
+                      min="0.1"
+                      placeholder="5.0"
                       {...field}
                       onChange={(e) => {
                         const value = e.target.value;
@@ -138,7 +153,6 @@ export default function ClientGoalForm({
                       value={field.value || ''}
                     />
                   </FormControl>
-                  <FormDescription>現在の目標: {currentGoal ? `${currentGoal}km` : '未設定'}</FormDescription>
                   <FormMessage />
                   {error && <p className="text-sm text-red-500">{error}</p>}
                 </FormItem>
@@ -149,13 +163,11 @@ export default function ClientGoalForm({
               <Button
                 type="submit"
                 disabled={isPending || form.formState.isSubmitting}
-                className="flex-1 bg-purple-500 hover:bg-purple-600"
+                className="flex-1 bg-emerald-500 hover:bg-emerald-600"
               >
                 {isPending || form.formState.isSubmitting
                   ? '保存中...'
-                  : showWelcomeMessage
-                    ? '目標を設定'
-                    : '目標を変更'}
+                  : '記録を保存'}
               </Button>
               <Button
                 type="button"
@@ -164,7 +176,7 @@ export default function ClientGoalForm({
                 disabled={isPending || form.formState.isSubmitting}
                 className="flex-1"
               >
-                {showWelcomeMessage ? '後で設定' : 'キャンセル'}
+                キャンセル
               </Button>
             </div>
           </form>
